Created embedded-graph-store & graph-store, and removed in-memory-store.

+++ firebase-store/src/main/kotlin/org/ostelco/prime/storage/firebase/FirebaseStorage.kt
@@ -32,10 +32,58 @@ class FirebaseStorage : Storage by FirebaseStorageSingleton
 
 object FirebaseStorageSingleton : Storage {
 
+    private val balanceEntity = EntityType("balance", Long::class.java)
+    private val productEntity = EntityType("products", Product::class.java)
+    private val subscriptionEntity = EntityType("subscriptions", String::class.java)
+    private val subscriberEntity = EntityType("subscribers", Subscriber::class.java)
+    private val paymentHistoryEntity = EntityType("paymentHistory", PurchaseRecord::class.java)
+    private val fcmTokenEntity = EntityType("notificationTokens", ApplicationToken::class.java)
+
+    private val firebaseDatabase = setupFirebaseInstance(config.databaseName, config.configFile)
+
+    private val balanceStore = EntityStore(firebaseDatabase, balanceEntity)
+    private val productStore = EntityStore(firebaseDatabase, productEntity)
+    private val subscriptionStore = EntityStore(firebaseDatabase, subscriptionEntity)
+    private val subscriberStore = EntityStore(firebaseDatabase, subscriberEntity)
+    private val paymentHistoryStore = EntityStore(firebaseDatabase, paymentHistoryEntity)
+    private val fcmTokenStore = EntityStore(firebaseDatabase, fcmTokenEntity)
+
+    private fun setupFirebaseInstance(
+            databaseName: String,
+            configFile: String): FirebaseDatabase {
+
+        try {
+
+            val credentials: GoogleCredentials = if (Files.exists(Paths.get(configFile))) {
+                FileInputStream(configFile).use { serviceAccount -> GoogleCredentials.fromStream(serviceAccount) }
+            } else {
+                GoogleCredentials.getApplicationDefault()
+            }
+
+            val options = FirebaseOptions.Builder()
+                    .setCredentials(credentials)
+                    .setDatabaseUrl("https://$databaseName.firebaseio.com/")
+                    .build()
+            try {
+                FirebaseApp.getInstance()
+            } catch (e: Exception) {
+                FirebaseApp.initializeApp(options)
+            }
+
+            return FirebaseDatabase.getInstance()
+
+            // (un)comment next line to turn on/of extended debugging
+            // from firebase.
+            // this.firebaseDatabase.setLogLevel(com.google.firebase.database.Logger.Level.DEBUG);
+        } catch (ex: IOException) {
+            throw StorageException(ex)
+        }
+    }
+
     override val balances: Map<String, Long>
         get() = balanceStore.getAll()
 
-    override fun addSubscriber(id: String, subscriber: Subscriber): Boolean = subscriberStore.create(id, subscriber)
+    override fun addSubscriber(subscriber: Subscriber) = subscriberStore.create(subscriber.id, subscriber)
 
     override fun getSubscriber(id: String): Subscriber? {
         val subscriber = subscriberStore.get(id)
@@ -43,20 +91,21 @@ object FirebaseStorageSingleton : Storage {
         return subscriber
     }
 
-    override fun updateSubscriber(id: String, subscriber: Subscriber): Boolean = subscriberStore.update(id, subscriber)
-
-    override fun getSubscription(id: String) = subscriptionStore.get(id)
+    override fun updateSubscriber(subscriber: Subscriber): Boolean = subscriberStore.update(subscriber.id, subscriber)
 
     override fun getMsisdn(subscriptionId: String) = subscriptionStore.get(subscriptionId)
 
-    override fun addSubscription(id: String, msisdn: String) {
-        subscriptionStore.create(id, msisdn)
-        balanceStore.create(msisdn, 0)
+    override fun addSubscription(id: String, msisdn: String): Boolean {
+        if (subscriptionStore.create(id, msisdn)) {
+            // should we set non-zero default balance?
+            return balanceStore.create(msisdn, 0)
+        }
+        return false
     }
 
-    override fun getProduct(sku: String) = productStore.get(sku)
+    override fun getProduct(subscriberId: String?, sku: String) = productStore.get(sku)
 
-    override fun getProducts() = productStore.getAll()
+    override fun getProducts(subscriberId: String): Map<String, Product> = productStore.getAll()
 
     override fun getBalance(id: String): Long? {
         val msisdn = subscriptionStore.get(id) ?: return null
@@ -79,15 +128,17 @@ object FirebaseStorageSingleton : Storage {
         }
     }
 
-    override fun removeSubscriber(id: String) {
+    override fun removeSubscriber(id: String): Boolean {
         subscriberStore.delete(id)
         // for payment history, skip checking if it exists.
         paymentHistoryStore.delete(id, dontExists = false)
         val msisdn = subscriptionStore.get(id)
         if (msisdn != null) {
-            subscriptionStore.delete(id)
-            balanceStore.delete(msisdn)
+            if (subscriptionStore.delete(id)) {
+                return balanceStore.delete(msisdn)
+            }
         }
+        return false
     }
 
     override fun addNotificationToken(msisdn: String, token: ApplicationToken) : Boolean {
@@ -105,54 +156,7 @@ object FirebaseStorageSingleton : Storage {
     }
 }
 
-val balanceEntity = EntityType("balance", Long::class.java)
-val productEntity = EntityType("products", Product::class.java)
-val subscriptionEntity = EntityType("subscriptions", String::class.java)
-val subscriberEntity = EntityType("subscribers", Subscriber::class.java)
-val paymentHistoryEntity = EntityType("paymentHistory", PurchaseRecord::class.java)
-val fcmTokenEntity = EntityType("notificationTokens", ApplicationToken::class.java)
-
-val config = FirebaseConfigRegistry.firebaseConfig
-val firebaseDatabase = setupFirebaseInstance(config.databaseName, config.configFile)
-
-val balanceStore = EntityStore(firebaseDatabase, balanceEntity)
-val productStore = EntityStore(firebaseDatabase, productEntity)
-val subscriptionStore = EntityStore(firebaseDatabase, subscriptionEntity)
-val subscriberStore = EntityStore(firebaseDatabase, subscriberEntity)
-val paymentHistoryStore = EntityStore(firebaseDatabase, paymentHistoryEntity)
-val fcmTokenStore = EntityStore(firebaseDatabase, fcmTokenEntity)
-
-private fun setupFirebaseInstance(
-        databaseName: String,
-        configFile: String): FirebaseDatabase {
-
-    try {
-
-        val credentials: GoogleCredentials = if (Files.exists(Paths.get(configFile))) {
-            FileInputStream(configFile).use { serviceAccount -> GoogleCredentials.fromStream(serviceAccount) }
-        } else {
-            GoogleCredentials.getApplicationDefault()
-        }
-
-        val options = FirebaseOptions.Builder()
-                .setCredentials(credentials)
-                .setDatabaseUrl("https://$databaseName.firebaseio.com/")
-                .build()
-        try {
-            FirebaseApp.getInstance()
-        } catch (e: Exception) {
-            FirebaseApp.initializeApp(options)
-        }
-
-        return FirebaseDatabase.getInstance()
-
-        // (un)comment next line to turn on/of extended debugging
-        // from firebase.
-        // this.firebaseDatabase.setLogLevel(com.google.firebase.database.Logger.Level.DEBUG);
-    } catch (ex: IOException) {
-        throw StorageException(ex)
-    }
-}
+private val config = FirebaseConfigRegistry.firebaseConfig
 
 const val TIMEOUT: Long = 10 //sec
